Make System.Data MONO-friendly
