+++ src/System.Data.Common/src/System/Data/Common/DbConnectionStringBuilder.cs
@@ -4,6 +4,7 @@
 
 using System.Collections;
 using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.Diagnostics;
 using System.Globalization;
@@ -237,7 +238,7 @@ namespace System.Data.Common
                     values[i] = this[keylist.Current];
                     Debug.Assert(null != values[i], "null value " + keylist.Current);
                 }
-                return new System.Data.Common.ReadOnlyCollection<object>(values);
+                return new ReadOnlyCollection<object>(values);
             }
         }
