+++ src/System.Data.Common/src/System/Data/Common/DbConnectionStringCommon.cs
@@ -4,86 +4,11 @@
 
 using System.Collections;
 using System.Collections.Generic;
-using System.ComponentModel;
 using System.Diagnostics;
 using System.Globalization;
 
 namespace System.Data.Common
 {
-    internal class DbConnectionStringBuilderDescriptor : PropertyDescriptor
-    {
-        internal DbConnectionStringBuilderDescriptor(string propertyName, Type componentType, Type propertyType, bool isReadOnly, Attribute[] attributes) : base(propertyName, attributes)
-        {
-            ComponentType = componentType;
-            PropertyType = propertyType;
-            IsReadOnly = isReadOnly;
-        }
-
-        internal bool RefreshOnChange { get; set; }
-        public override Type ComponentType { get; }
-        public override bool IsReadOnly { get; }
-        public override Type PropertyType { get; }
-
-        public override bool CanResetValue(object component)
-        {
-            DbConnectionStringBuilder builder = (component as DbConnectionStringBuilder);
-            return ((null != builder) && builder.ShouldSerialize(DisplayName));
-        }
-
-        public override object GetValue(object component)
-        {
-            DbConnectionStringBuilder builder = (component as DbConnectionStringBuilder);
-            if (null != builder)
-            {
-                object value;
-                if (builder.TryGetValue(DisplayName, out value))
-                {
-                    return value;
-                }
-            }
-            return null;
-        }
-
-        public override void ResetValue(object component)
-        {
-            DbConnectionStringBuilder builder = (component as DbConnectionStringBuilder);
-            if (null != builder)
-            {
-                builder.Remove(DisplayName);
-
-                if (RefreshOnChange)
-                {
-                    builder.ClearPropertyDescriptors();
-                }
-            }
-        }
-
-        public override void SetValue(object component, object value)
-        {
-            DbConnectionStringBuilder builder = (component as DbConnectionStringBuilder);
-            if (null != builder)
-            {
-                // via the editor, empty string does a defacto Reset
-                if ((typeof(string) == PropertyType) && string.Empty.Equals(value))
-                {
-                    value = null;
-                }
-                builder[DisplayName] = value;
-
-                if (RefreshOnChange)
-                {
-                    builder.ClearPropertyDescriptors();
-                }
-            }
-        }
-
-        public override bool ShouldSerializeValue(object component)
-        {
-            DbConnectionStringBuilder builder = (component as DbConnectionStringBuilder);
-            return ((null != builder) && builder.ShouldSerialize(DisplayName));
-        }
-    }
-
     [Serializable]
     internal sealed class ReadOnlyCollection<T> : ICollection, ICollection<T>
     {
