+++ src/System.Data.SqlClient/src/System/Data/SqlClient/TdsParser.cs
@@ -40,7 +40,9 @@ namespace System.Data.SqlClient
             // For a default installation of SqlServer the encoding exchanged during Login is 1252. This encoding is not loaded by default
             // See Remarks at https://msdn.microsoft.com/en-us/library/system.text.encodingprovider(v=vs.110).aspx 
             // SqlClient needs to register the encoding providers to make sure that even basic scenarios work with Sql Server.
+#if !MONO
             Encoding.RegisterProvider(CodePagesEncodingProvider.Instance);
+#endif
         }
 
         // Default state object for parser
